Detect and handle reorgs in sync package
Reorgs are detected when blockchain gets a block at expected height
but it doesnt have the expected parent block hash. In that case,
synchronizer will revert the HEAD block and retry to start syncing from
that point on again. If the reorg is deeper than a single block, synchronizer
will keep reverting the HEAD and retry.